requirements.txt
aiohttp~=3.8.4
aiohttp_jinja2~=1.5.1
arkprts==0.2.1
beautifulsoup4~=4.12.2

python-dotenv~=1.0.0
